added a flag to have large table stories disabled by default. in the process, story generation for SummaryView and DetailView tables has been moved to a shared module so they can share that logic and reduce code duplication.
