+++ apps/src/templates/sectionProgress/progressTables/ProgressTableDetailView.story.jsx
@@ -1,75 +1,15 @@
-import React from 'react';
-import ProgressTableDetailView from '@cdo/apps/templates/sectionProgress/progressTables/ProgressTableDetailView';
-import {Provider} from 'react-redux';
-import {createStore, wrapTable} from '../sectionProgressTestHelpers';
+import {detailTableStories} from '@cdo/apps/templates/sectionProgress/progressTables/sectionProgressTableStoryBuilder';
 
+/**
+ * Stories for the SummaryView and DetailView are identical other than the
+ * actual component to render, so we build them using common code in
+ * `sectionProgressTableStoryBuilder`.
+ *
+ * Note: there is a flag in that file to enable additional stories that is
+ * disabled by default to prevent slowing down our test pipeline.
+ */
 export default storybook => {
-  storybook.storiesOf('SectionProgress/DetailView', module).addStoryTable([
-    {
-      name: 'Tiny section, small script',
-      story: () => {
-        const store = createStore(3, 10);
-        return wrapTable(
-          <Provider store={store}>
-            <ProgressTableDetailView />
-          </Provider>
-        );
-      }
-    },
-    {
-      name: 'Tiny section, large script',
-      story: () => {
-        const store = createStore(3, 30);
-        return wrapTable(
-          <Provider store={store}>
-            <ProgressTableDetailView />
-          </Provider>
-        );
-      }
-    },
-    {
-      name: 'Small section, small script',
-      story: () => {
-        const store = createStore(30, 10);
-        return wrapTable(
-          <Provider store={store}>
-            <ProgressTableDetailView />
-          </Provider>
-        );
-      }
-    },
-    {
-      name: 'Small section, large script',
-      story: () => {
-        const store = createStore(30, 30);
-        return wrapTable(
-          <Provider store={store}>
-            <ProgressTableDetailView />
-          </Provider>
-        );
-      }
-    },
-    {
-      name: 'Large section, small script',
-      story: () => {
-        const store = createStore(200, 10);
-        return wrapTable(
-          <Provider store={store}>
-            <ProgressTableDetailView />
-          </Provider>
-        );
-      }
-    },
-    {
-      name: 'Large section, large script',
-      story: () => {
-        const store = createStore(200, 30);
-        return wrapTable(
-          <Provider store={store}>
-            <ProgressTableDetailView />
-          </Provider>
-        );
-      }
-    }
-  ]);
+  storybook
+    .storiesOf('SectionProgress/DetailView', module)
+    .addStoryTable(detailTableStories);
 };
